Refactor all methods that scrape data from youtube into a new module
