+++ web/talks/models.py
@@ -1,14 +1,8 @@
-import requests
-
 from datetime import timedelta
-
-from urllib.parse import urlsplit
-from urllib.parse import parse_qs
 
 from django.db import models
 from django.utils import timezone
 from django.utils.text import slugify
-from django.core.management.base import CommandError
 
 from taggit.managers import TaggableManager
 
@@ -139,87 +133,3 @@ class Talk(models.Model):
         verbose_name_plural = "Talks"
         get_latest_by = "-created"
         ordering = ['-created', '-updated']
-
-
-def get_video_code(url):
-    query = urlsplit(url).query
-    params = parse_qs(query)
-    if "v" not in params:
-        raise CommandError('Invalid url "%s"' % url)
-    video_code = params["v"][0]
-    return video_code
-
-
-def get_playlist_code(url):
-    query = urlsplit(url).query
-    params = parse_qs(query)
-    if "list" not in params:
-        raise CommandError('Invalid url "%s"' % url)
-    playlist_code = params["list"][0]
-    return playlist_code
-
-
-def fetch_channel_data(youtube_api_key, channel_code):
-    channel_url = "https://www.googleapis.com/youtube/v3/channels"
-    payload = {'id': channel_code,
-               'part': 'snippet,contentDetails',
-               'key': youtube_api_key}
-    resp = requests.get(channel_url, params=payload)
-    if resp.status_code != 200:
-        raise CommandError('Error fetching channel data "%s"' % resp.status_code)
-    response_json = resp.json()
-    channel_data = None
-    if len(response_json["items"]) > 0:
-        channel_data = response_json["items"][0]
-    return channel_data
-
-
-def fetch_playlist_data(youtube_api_key, playlist_code):
-    channel_url = "https://www.googleapis.com/youtube/v3/playlists"
-    payload = {'id': playlist_code,
-               'part': 'snippet,contentDetails',
-               'key': youtube_api_key}
-    resp = requests.get(channel_url, params=payload)
-    if resp.status_code != 200:
-        raise CommandError(
-            'Error fetching playlist data "%s"' % resp.status_code)
-    response_json = resp.json()
-    channel_data = None
-    if len(response_json["items"]) > 0:
-        channel_data = response_json["items"][0]
-    return channel_data
-
-
-def fetch_video_data(youtube_api_key, video_code):
-    video_url = "https://www.googleapis.com/youtube/v3/videos"
-    payload = {'id': video_code,
-               'part': 'snippet,contentDetails,statistics,topicDetails,status,recordingDetails,player,localizations,liveStreamingDetails',
-               'key': youtube_api_key}
-    resp = requests.get(video_url, params=payload)
-    if resp.status_code != 200:
-        raise CommandError('Error fetching video data "%s"' % resp.status_code)
-    response_json = resp.json()
-    video_data = None
-    if len(response_json["items"]) > 0:
-        video_data = response_json["items"][0]
-    return video_data
-
-
-def fetch_playlist_items(youtube_api_key, playlist_code):
-    channel_url = "https://www.googleapis.com/youtube/v3/playlistItems"
-    payload = {'playlistId': playlist_code,
-               'maxResults': 50,
-               'part': 'snippet',
-               'key': youtube_api_key}
-    resp = requests.get(channel_url, params=payload)
-    if resp.status_code != 200:
-        if resp.status_code == 404:
-            return []
-        else:
-            raise CommandError('Error fetching playlist items "%s"' % resp.status_code)
-    response_json = resp.json()
-    videos_id_list = []
-    data_json = response_json["items"]
-    for item in data_json:
-        videos_id_list.append(item["snippet"]["resourceId"]["videoId"])
-    return videos_id_list
